+++ bin/release_packer.dart
@@ -5,9 +5,24 @@ import 'package:release_updater/release_packer.dart';
 import 'package:release_updater/release_updater.dart';
 import 'package:release_updater/src/release_updater_config.dart';
 
+const _hr1 =
+    '══════════════════════════════════════════════════════════════════════════════';
+
+const _hr2 =
+    '──────────────────────────────────────────────────────────────────────────────';
+
 void main(List<String> args) async {
-  print('--------------------------------------------------------------------');
-  print('[ release_packer/${ReleaseUpdater.VERSION} ]\n');
+  var dartVersion = Platform.version.split(' ')[0];
+  var operatingSystem = Platform.operatingSystem;
+
+  var title =
+      'release_packer/${ReleaseUpdater.VERSION} | Dart $dartVersion ($operatingSystem)';
+
+  print('''
+╔══${'═' * title.length}${'═' * (_hr1.length - title.length - 6)}══╗
+║  $title ${' ' * (_hr1.length - title.length - 6)} ║
+╚══${'═' * title.length}${'═' * (_hr1.length - title.length - 6)}══╝
+''');
 
   if (args.length < 2) {
     print('USAGE:\n');
@@ -27,7 +42,9 @@ void main(List<String> args) async {
 
   var cmd = args.removeAt(0).toLowerCase();
 
-  print('-- Loading `ReleasePacker` from: $releasePackerJsonPath');
+  print('»  Current directory: ${Directory.current.path}');
+
+  print('»  Loading `ReleasePacker` from: $releasePackerJsonPath');
   var releasePacker =
       ReleasePacker.fromFilePath(releasePackerJsonPath, properties: properties);
 
@@ -61,49 +78,49 @@ void main(List<String> args) async {
 
     var releaseZipFile = File(releaseZipPath).absolute;
 
-    print('-- Generating Zip...');
+    print('»  Generating Release Zip...');
     var zipBytes = await releaseBundle.zipBytes;
     releaseZipFile.writeAsBytesSync(zipBytes);
 
-    print('-- Saved ${releaseZipFile.lengthSync()} bytes.');
+    print('»  Release Zip size: ${releaseZipFile.lengthSync()} bytes.');
 
     print(
-        '-- Release `${releaseBundle.release}` at:\n\n  ${releaseZipFile.path}\n');
+        '\n»  Release `${releaseBundle.release}` saved at:\n\n  ${releaseZipFile.path}\n');
   } else {
-    print('** Unknown command: $cmd $args\n');
+    print('▒  Unknown command: $cmd $args\n');
     exit(1);
   }
 
-  print('--------------------------------------------------------------------');
+  print(_hr1);
   exit(0);
 }
 
 void _showReleasePacker(ReleasePacker releasePacker, {bool showFiles = false}) {
-  print('\n-- Release name: ${releasePacker.name}');
-  print('-- Release version: ${releasePacker.version}');
+  print('\n»  Release name: ${releasePacker.name}');
+  print('»  Release version: ${releasePacker.version}');
 
   var prepareCommands = releasePacker.prepareCommands;
   if (prepareCommands != null) {
-    print('\n-- Prepare commands (${prepareCommands.length}):');
+    print('\n»  Prepare commands (${prepareCommands.length}):');
     for (var cmd in prepareCommands) {
-      print('  -- $cmd');
+      print('   -  $cmd');
     }
   }
 
   var finalizeCommands = releasePacker.finalizeCommands;
   if (finalizeCommands != null) {
-    print('\n-- Finalize commands (${finalizeCommands.length}):');
+    print('\n»  Finalize commands (${finalizeCommands.length}):');
     for (var cmd in finalizeCommands) {
-      print('  -- $cmd');
+      print('   -  $cmd');
     }
   }
 
   if (showFiles) {
     var files = releasePacker.files;
-    print('\n-- Files ${files.length}:');
+    print('\n»  Files ${files.length}:');
 
     for (var f in files) {
-      print('  -- $f');
+      print('   »  $f');
     }
   }
 }
@@ -111,29 +128,29 @@ void _showReleasePacker(ReleasePacker releasePacker, {bool showFiles = false}) {
 Future<ReleaseBundleZip> _buildReleaseBundle(
     ReleasePacker releasePacker, String sourcePath,
     [String? releasesPath]) async {
+  print('\n$_hr2');
+
+  print('\n»»  Generating release bundle...\n');
+
   print(
-      '\n--------------------------------------------------------------------');
-  print(
-      '** Building release: $sourcePath${releasesPath != null ? ' -> $releasesPath' : ''}\n');
+      '» Building release from `$sourcePath`${releasesPath != null ? ' to `$releasesPath`' : ''}');
 
   var platform = ReleasePlatform.platform;
-
-  print('-- Generating release bundle...');
 
   var releaseBundle = await releasePacker.buildFromDirectory(
       sourcePath: sourcePath, platform: platform);
 
-  print('-- Generated release: ${releaseBundle.release}');
+  print('»  Generated release: `${releaseBundle.release}`');
   return releaseBundle;
 }
 
 Future<void> _showBundleFiles(ReleaseBundleZip releaseBundle) async {
   var files = await releaseBundle.files;
 
-  print('\n-- Bundle files (${files.length}):');
-  for (var f in files) {
-    var execStr = f.executable ? ' (EXEC)' : '';
-    print('  -- ${f.filePath} (${await f.length} bytes)$execStr');
-  }
-  print('');
+  print('\n»  Bundle files (${files.length}):');
+
+  var asciiArtTree = await releaseBundle.toASCIIArtTree();
+  var tree = asciiArtTree.generate(indent: '   ║  ', trim: true);
+
+  print('\n$tree\n');
 }